Fix models for migration generation Adds a new version of the relationship and column overrides to account for the use of Field() and Relationship() on their own, and updates some types of fields to use sa_type instead of sa_column in order to get to server to stop generating the wrong kind of column in migrations. Also removes datetime.utcnow() as it's deprecated.
