+++ alembic/versions/b34f03d164c6_relationships_overhaul.py
@@ -1,0 +1,630 @@
+"""Relationships overhaul
+
+Revision ID: b34f03d164c6
+Revises: 3d942ce689fc
+Create Date: 2026-02-19 16:25:44.877345
+
+"""
+
+
+# revision identifiers, used by Alembic.
+revision = 'b34f03d164c6'
+down_revision = '3d942ce689fc'
+branch_labels = None
+depends_on = None
+
+from alembic import op
+import sqlalchemy as sa
+from sqlalchemy.dialects import postgresql
+
+
+try:
+    is_sqlite = op.get_context().dialect.name == 'sqlite'
+except Exception:
+    is_sqlite = False
+
+if is_sqlite:
+    op.get_context().connection.execute('PRAGMA foreign_keys=ON;')
+    utcnow_server_default = "(datetime('now', 'utc'))"
+else:
+    utcnow_server_default = "timezone('utc', current_timestamp)"
+
+def sqlite_column_reflect_listener(inspector, table, column_info):
+    """Adds parenthesis around SQLite datetime defaults for utcnow."""
+    if column_info['default'] == "datetime('now', 'utc')":
+        column_info['default'] = utcnow_server_default
+
+sqlite_reflect_kwargs = {
+    'listeners': [('column_reflect', sqlite_column_reflect_listener)]
+}
+
+# ===========================================================================
+# HOWTO: Handle alter statements in SQLite
+#
+# def upgrade():
+#     if is_sqlite:
+#         with op.batch_alter_table('table_name', reflect_kwargs=sqlite_reflect_kwargs) as batch_op:
+#             batch_op.alter_column('column_name', type_=sa.Unicode(), server_default='', nullable=False)
+#     else:
+#         op.alter_column('table_name', 'column_name', type_=sa.Unicode(), server_default='', nullable=False)
+#
+# ===========================================================================
+
+
+def upgrade():
+    op.drop_table('attendee_tournament')
+    op.drop_constraint(op.f('fk_admin_account_attendee_id_attendee'), 'admin_account', type_='foreignkey')
+    op.create_foreign_key(None, 'admin_account', 'attendee', ['attendee_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_api_token_admin_account_id_admin_account'), 'api_token', type_='foreignkey')
+    op.create_foreign_key(None, 'api_token', 'admin_account', ['admin_account_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_art_panel_assignment_panel_id_art_show_panel'), 'art_panel_assignment', type_='foreignkey')
+    op.drop_constraint(op.f('fk_art_panel_assignment_app_id_art_show_application'), 'art_panel_assignment', type_='foreignkey')
+    op.create_foreign_key(None, 'art_panel_assignment', 'art_show_application', ['app_id'], ['id'], ondelete='CASCADE')
+    op.create_foreign_key(None, 'art_panel_assignment', 'art_show_panel', ['panel_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_art_show_agent_code_attendee_id_attendee'), 'art_show_agent_code', type_='foreignkey')
+    op.drop_constraint(op.f('fk_art_show_agent_code_app_id_art_show_application'), 'art_show_agent_code', type_='foreignkey')
+    op.create_foreign_key(None, 'art_show_agent_code', 'art_show_application', ['app_id'], ['id'], ondelete='CASCADE')
+    op.create_foreign_key(None, 'art_show_agent_code', 'attendee', ['attendee_id'], ['id'])
+    op.alter_column('art_show_application', 'overridden_price',
+               existing_type=sa.INTEGER(),
+               nullable=False)
+    op.alter_column('art_show_application', 'default_cost',
+               existing_type=sa.INTEGER(),
+               nullable=False)
+    op.create_unique_constraint(None, 'art_show_application', ['attendee_id'])
+    op.drop_constraint(op.f('fk_art_show_application_attendee_id_attendee'), 'art_show_application', type_='foreignkey')
+    op.create_foreign_key(None, 'art_show_application', 'attendee', ['attendee_id'], ['id'])
+    op.alter_column('art_show_bidder', 'attendee_id',
+               existing_type=sa.UUID(),
+               nullable=False)
+    op.drop_constraint(op.f('fk_art_show_bidder_attendee_id_attendee'), 'art_show_bidder', type_='foreignkey')
+    op.create_foreign_key(None, 'art_show_bidder', 'attendee', ['attendee_id'], ['id'], ondelete='CASCADE')
+    op.alter_column('art_show_payment', 'receipt_id',
+               existing_type=sa.UUID(),
+               nullable=False)
+    op.drop_constraint(op.f('fk_art_show_payment_receipt_id_art_show_receipt'), 'art_show_payment', type_='foreignkey')
+    op.create_foreign_key(None, 'art_show_payment', 'art_show_receipt', ['receipt_id'], ['id'], ondelete='CASCADE')
+    op.alter_column('art_show_piece', 'app_id',
+               existing_type=sa.UUID(),
+               nullable=False)
+    op.drop_constraint(op.f('fk_art_show_piece_receipt_id_art_show_receipt'), 'art_show_piece', type_='foreignkey')
+    op.drop_constraint(op.f('fk_art_show_piece_winning_bidder_id_art_show_bidder'), 'art_show_piece', type_='foreignkey')
+    op.drop_constraint(op.f('fk_art_show_pieces_app_id_art_show_application'), 'art_show_piece', type_='foreignkey')
+    op.create_foreign_key(None, 'art_show_piece', 'art_show_receipt', ['receipt_id'], ['id'])
+    op.create_foreign_key(None, 'art_show_piece', 'art_show_bidder', ['winning_bidder_id'], ['id'])
+    op.create_foreign_key(None, 'art_show_piece', 'art_show_application', ['app_id'], ['id'], ondelete='CASCADE')
+    op.alter_column('art_show_receipt', 'attendee_id',
+               existing_type=sa.UUID(),
+               nullable=False)
+    op.drop_constraint(op.f('fk_art_show_receipt_attendee_id_attendee'), 'art_show_receipt', type_='foreignkey')
+    op.create_foreign_key(None, 'art_show_receipt', 'attendee', ['attendee_id'], ['id'], ondelete='CASCADE')
+    op.create_unique_constraint(None, 'artist_marketplace_application', ['attendee_id'])
+    op.drop_constraint(op.f('fk_artist_marketplace_application_attendee_id_attendee'), 'artist_marketplace_application', type_='foreignkey')
+    op.create_foreign_key(None, 'artist_marketplace_application', 'attendee', ['attendee_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_attendee_creator_id_attendee'), 'attendee', type_='foreignkey')
+    op.drop_constraint(op.f('fk_attendee_watchlist_id_watch_list'), 'attendee', type_='foreignkey')
+    op.drop_constraint(op.f('fk_attendee_group_id_group'), 'attendee', type_='foreignkey')
+    op.drop_constraint(op.f('fk_attendee_badge_pickup_group_id_badge_pickup_group'), 'attendee', type_='foreignkey')
+    op.create_foreign_key(None, 'attendee', 'attendee', ['creator_id'], ['id'])
+    op.create_foreign_key(None, 'attendee', 'watch_list', ['watchlist_id'], ['id'])
+    op.create_foreign_key(None, 'attendee', 'group', ['group_id'], ['id'])
+    op.create_foreign_key(None, 'attendee', 'badge_pickup_group', ['badge_pickup_group_id'], ['id'])
+    op.drop_column('attendee', 'affiliate')
+    op.drop_constraint(op.f('attraction_name_key'), 'attraction', type_='unique')
+    op.drop_constraint(op.f('uq_attraction_slug'), 'attraction', type_='unique')
+    op.drop_constraint(op.f('fk_attraction_event_attraction_feature_id_attraction_feature'), 'attraction_event', type_='foreignkey')
+    op.drop_constraint(op.f('fk_attraction_event_attraction_id_attraction'), 'attraction_event', type_='foreignkey')
+    op.drop_constraint(op.f('fk_attraction_event_location_id_event_location'), 'attraction_event', type_='foreignkey')
+    op.create_foreign_key(None, 'attraction_event', 'attraction_feature', ['attraction_feature_id'], ['id'], ondelete='CASCADE')
+    op.create_foreign_key(None, 'attraction_event', 'attraction', ['attraction_id'], ['id'], ondelete='CASCADE')
+    op.create_foreign_key(None, 'attraction_event', 'event_location', ['event_location_id'], ['id'])
+    op.drop_constraint(op.f('fk_attraction_feature_attraction_id_attraction'), 'attraction_feature', type_='foreignkey')
+    op.create_foreign_key(None, 'attraction_feature', 'attraction', ['attraction_id'], ['id'], ondelete='CASCADE')
+    op.drop_index(op.f('ix_attraction_notification_ident'), table_name='attraction_notification')
+    op.drop_constraint(op.f('fk_attraction_notification_attraction_id_attraction'), 'attraction_notification', type_='foreignkey')
+    op.drop_constraint(op.f('fk_attraction_notification_attraction_event_id_attraction_event'), 'attraction_notification', type_='foreignkey')
+    op.drop_constraint(op.f('fk_attraction_notification_attendee_id_attendee'), 'attraction_notification', type_='foreignkey')
+    op.create_foreign_key(None, 'attraction_notification', 'attraction_event', ['attraction_event_id'], ['id'], ondelete='CASCADE')
+    op.create_foreign_key(None, 'attraction_notification', 'attendee', ['attendee_id'], ['id'], ondelete='CASCADE')
+    op.create_foreign_key(None, 'attraction_notification', 'attraction', ['attraction_id'], ['id'], ondelete='CASCADE')
+    op.drop_index(op.f('ix_attraction_notification_reply_sid'), table_name='attraction_notification_reply')
+    op.drop_constraint(op.f('fk_attraction_signup_attendee_id_attendee'), 'attraction_signup', type_='foreignkey')
+    op.drop_constraint(op.f('fk_attraction_signup_attraction_id_attraction'), 'attraction_signup', type_='foreignkey')
+    op.drop_constraint(op.f('fk_attraction_signup_attraction_event_id_attraction_event'), 'attraction_signup', type_='foreignkey')
+    op.create_foreign_key(None, 'attraction_signup', 'attendee', ['attendee_id'], ['id'], ondelete='CASCADE')
+    op.create_foreign_key(None, 'attraction_signup', 'attraction', ['attraction_id'], ['id'], ondelete='CASCADE')
+    op.create_foreign_key(None, 'attraction_signup', 'attraction_event', ['attraction_event_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('uq_automated_email_ident'), 'automated_email', type_='unique')
+    op.drop_index(op.f('ix_badge_info_ident'), table_name='badge_info')
+    op.drop_constraint(op.f('fk_badge_info_attendee_id_attendee'), 'badge_info', type_='foreignkey')
+    op.create_foreign_key(None, 'badge_info', 'attendee', ['attendee_id'], ['id'])
+    op.drop_constraint(op.f('fk_bulk_printing_request_department_id_department'), 'bulk_printing_request', type_='foreignkey')
+    op.create_foreign_key(None, 'bulk_printing_request', 'department', ['department_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_department_parent_id_department'), 'department', type_='foreignkey')
+    op.create_foreign_key(None, 'department', 'department', ['parent_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_dept_checklist_item_department_id_department'), 'dept_checklist_item', type_='foreignkey')
+    op.drop_constraint(op.f('fk_dept_checklist_item_attendee_id_attendee'), 'dept_checklist_item', type_='foreignkey')
+    op.create_foreign_key(None, 'dept_checklist_item', 'department', ['department_id'], ['id'], ondelete='CASCADE')
+    op.create_foreign_key(None, 'dept_checklist_item', 'attendee', ['attendee_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_dept_membership_department_id_department'), 'dept_membership', type_='foreignkey')
+    op.drop_constraint(op.f('fk_dept_membership_attendee_id_attendee'), 'dept_membership', type_='foreignkey')
+    op.create_foreign_key(None, 'dept_membership', 'attendee', ['attendee_id'], ['id'], ondelete='CASCADE')
+    op.create_foreign_key(None, 'dept_membership', 'department', ['department_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_dept_membership_request_attendee_id_attendee'), 'dept_membership_request', type_='foreignkey')
+    op.drop_constraint(op.f('fk_dept_membership_request_department_id_department'), 'dept_membership_request', type_='foreignkey')
+    op.create_foreign_key(None, 'dept_membership_request', 'attendee', ['attendee_id'], ['id'], ondelete='CASCADE')
+    op.create_foreign_key(None, 'dept_membership_request', 'department', ['department_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_dept_role_department_id_department'), 'dept_role', type_='foreignkey')
+    op.create_foreign_key(None, 'dept_role', 'department', ['department_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_email_automated_email_id_automated_email'), 'email', type_='foreignkey')
+    op.create_foreign_key(None, 'email', 'automated_email', ['automated_email_id'], ['id'])
+    op.create_unique_constraint(None, 'event', ['attraction_event_id'])
+    op.drop_constraint(op.f('fk_event_attraction_event_id_attraction_event'), 'event', type_='foreignkey')
+    op.drop_constraint(op.f('fk_event_department_id_department'), 'event', type_='foreignkey')
+    op.drop_constraint(op.f('fk_event_location_id_event_location'), 'event', type_='foreignkey')
+    op.create_foreign_key(None, 'event', 'attraction_event', ['attraction_event_id'], ['id'])
+    op.create_foreign_key(None, 'event', 'department', ['department_id'], ['id'])
+    op.create_foreign_key(None, 'event', 'event_location', ['event_location_id'], ['id'])
+    op.drop_constraint(op.f('fk_event_feedback_event_id_event'), 'event_feedback', type_='foreignkey')
+    op.create_foreign_key(None, 'event_feedback', 'event', ['event_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_event_location_department_id_department'), 'event_location', type_='foreignkey')
+    op.create_foreign_key(None, 'event_location', 'department', ['department_id'], ['id'])
+    op.drop_constraint(op.f('fk_food_restrictions_attendee_id_attendee'), 'food_restrictions', type_='foreignkey')
+    op.create_foreign_key(None, 'food_restrictions', 'attendee', ['attendee_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_leader'), 'group', type_='foreignkey')
+    op.drop_constraint(op.f('fk_group_shared_with_id_group'), 'group', type_='foreignkey')
+    op.create_foreign_key(None, 'group', 'attendee', ['leader_id'], ['id'])
+    op.create_foreign_key(None, 'group', 'group', ['shared_with_id'], ['id'])
+    op.drop_constraint(op.f('fk_guest_autograph_guest_id_guest_group'), 'guest_autograph', type_='foreignkey')
+    op.create_foreign_key(None, 'guest_autograph', 'guest_group', ['guest_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_guest_bio_guest_id_guest_group'), 'guest_bio', type_='foreignkey')
+    op.create_foreign_key(None, 'guest_bio', 'guest_group', ['guest_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_guest_charity_guest_id_guest_group'), 'guest_charity', type_='foreignkey')
+    op.create_foreign_key(None, 'guest_charity', 'guest_group', ['guest_id'], ['id'], ondelete='CASCADE')
+    op.alter_column('guest_detailed_travel_plan', 'travel_plans_id',
+               existing_type=sa.UUID(),
+               nullable=False)
+    op.drop_constraint(op.f('fk_guest_detailed_travel_plan_travel_plans_id_guest_tra_6ad4'), 'guest_detailed_travel_plan', type_='foreignkey')
+    op.create_foreign_key(None, 'guest_detailed_travel_plan', 'guest_travel_plans', ['travel_plans_id'], ['id'], ondelete='CASCADE')
+    op.create_unique_constraint(None, 'guest_group', ['group_id'])
+    op.drop_constraint(op.f('fk_guest_group_group_id_group'), 'guest_group', type_='foreignkey')
+    op.drop_constraint(op.f('fk_guest_group_event_id_event'), 'guest_group', type_='foreignkey')
+    op.create_foreign_key(None, 'guest_group', 'group', ['group_id'], ['id'], ondelete='CASCADE')
+    op.create_foreign_key(None, 'guest_group', 'event', ['event_id'], ['id'])
+    op.drop_constraint(op.f('fk_guest_hospitality_guest_id_guest_group'), 'guest_hospitality', type_='foreignkey')
+    op.create_foreign_key(None, 'guest_hospitality', 'guest_group', ['guest_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_guest_image_guest_id_guest_group'), 'guest_image', type_='foreignkey')
+    op.create_foreign_key(None, 'guest_image', 'guest_group', ['guest_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_guest_info_guest_id_guest_group'), 'guest_info', type_='foreignkey')
+    op.create_foreign_key(None, 'guest_info', 'guest_group', ['guest_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_guest_interview_guest_id_guest_group'), 'guest_interview', type_='foreignkey')
+    op.create_foreign_key(None, 'guest_interview', 'guest_group', ['guest_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_guest_media_request_guest_id_guest_group'), 'guest_media_request', type_='foreignkey')
+    op.create_foreign_key(None, 'guest_media_request', 'guest_group', ['guest_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_guest_merch_guest_id_guest_group'), 'guest_merch', type_='foreignkey')
+    op.create_foreign_key(None, 'guest_merch', 'guest_group', ['guest_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_guest_panel_guest_id_guest_group'), 'guest_panel', type_='foreignkey')
+    op.create_foreign_key(None, 'guest_panel', 'guest_group', ['guest_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_guest_stage_plot_guest_id_guest_group'), 'guest_stage_plot', type_='foreignkey')
+    op.create_foreign_key(None, 'guest_stage_plot', 'guest_group', ['guest_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_guest_taxes_guest_id_guest_group'), 'guest_taxes', type_='foreignkey')
+    op.create_foreign_key(None, 'guest_taxes', 'guest_group', ['guest_id'], ['id'], ondelete='CASCADE')
+    op.create_unique_constraint(None, 'guest_track', ['guest_id'])
+    op.drop_constraint(op.f('fk_guest_track_guest_id_guest_group'), 'guest_track', type_='foreignkey')
+    op.create_foreign_key(None, 'guest_track', 'guest_group', ['guest_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_guest_travel_plans_guest_id_guest_group'), 'guest_travel_plans', type_='foreignkey')
+    op.create_foreign_key(None, 'guest_travel_plans', 'guest_group', ['guest_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_hotel_requests_attendee_id_attendee'), 'hotel_requests', type_='foreignkey')
+    op.create_foreign_key(None, 'hotel_requests', 'attendee', ['attendee_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_indie_developer_studio_id_indie_studio'), 'indie_developer', type_='foreignkey')
+    op.create_foreign_key(None, 'indie_developer', 'indie_studio', ['studio_id'], ['id'], ondelete='CASCADE')
+    op.alter_column('indie_game', 'multiplayer_game_length',
+               existing_type=sa.INTEGER(),
+               nullable=False)
+    op.drop_constraint(op.f('fk_indie_game_primary_contact_id_indie_developer'), 'indie_game', type_='foreignkey')
+    op.drop_constraint(op.f('fk_indie_game_studio_id_indie_studio'), 'indie_game', type_='foreignkey')
+    op.create_foreign_key(None, 'indie_game', 'indie_developer', ['primary_contact_id'], ['id'])
+    op.create_foreign_key(None, 'indie_game', 'indie_studio', ['studio_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_indie_game_code_game_id_indie_game'), 'indie_game_code', type_='foreignkey')
+    op.create_foreign_key(None, 'indie_game_code', 'indie_game', ['game_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_indie_game_image_game_id_indie_game'), 'indie_game_image', type_='foreignkey')
+    op.create_foreign_key(None, 'indie_game_image', 'indie_game', ['game_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_indie_game_review_game_id_indie_game'), 'indie_game_review', type_='foreignkey')
+    op.drop_constraint(op.f('fk_indie_game_review_judge_id_indie_judge'), 'indie_game_review', type_='foreignkey')
+    op.create_foreign_key(None, 'indie_game_review', 'indie_game', ['game_id'], ['id'], ondelete='CASCADE')
+    op.create_foreign_key(None, 'indie_game_review', 'indie_judge', ['judge_id'], ['id'], ondelete='CASCADE')
+    op.alter_column('indie_judge', 'admin_id',
+               existing_type=sa.UUID(),
+               nullable=True)
+    op.create_unique_constraint(None, 'indie_judge', ['admin_id'])
+    op.drop_constraint(op.f('uq_indie_studio_name'), 'indie_studio', type_='unique')
+    op.create_unique_constraint(None, 'indie_studio', ['group_id'])
+    op.drop_constraint(op.f('fk_job_department_id_department'), 'job', type_='foreignkey')
+    op.create_foreign_key(None, 'job', 'department', ['department_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_job_template_department_id_department'), 'job_template', type_='foreignkey')
+    op.create_foreign_key(None, 'job_template', 'department', ['department_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('uq_lottery_application_response_id'), 'lottery_application', type_='unique')
+    op.drop_constraint(op.f('fk_m_points_for_cash_attendee_id_attendee'), 'm_points_for_cash', type_='foreignkey')
+    op.create_foreign_key(None, 'm_points_for_cash', 'attendee', ['attendee_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_merch_discount_attendee_id_attendee'), 'merch_discount', type_='foreignkey')
+    op.create_foreign_key(None, 'merch_discount', 'attendee', ['attendee_id'], ['id'], ondelete='CASCADE')
+    op.alter_column('merch_pickup', 'picked_up_by_id',
+               existing_type=sa.UUID(),
+               nullable=True)
+    op.alter_column('merch_pickup', 'picked_up_for_id',
+               existing_type=sa.UUID(),
+               nullable=True)
+    op.drop_constraint(op.f('fk_mits_applicant_team_id_mits_team'), 'mits_applicant', type_='foreignkey')
+    op.create_foreign_key(None, 'mits_applicant', 'mits_team', ['team_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_mits_document_game_id_mits_game'), 'mits_document', type_='foreignkey')
+    op.create_foreign_key(None, 'mits_document', 'mits_game', ['game_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_mits_game_team_id_mits_team'), 'mits_game', type_='foreignkey')
+    op.create_foreign_key(None, 'mits_game', 'mits_team', ['team_id'], ['id'], ondelete='CASCADE')
+    op.create_unique_constraint(None, 'mits_panel_application', ['team_id'])
+    op.drop_constraint(op.f('fk_mits_panel_application_team_id_mits_team'), 'mits_panel_application', type_='foreignkey')
+    op.create_foreign_key(None, 'mits_panel_application', 'mits_team', ['team_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_mits_picture_game_id_mits_game'), 'mits_picture', type_='foreignkey')
+    op.create_foreign_key(None, 'mits_picture', 'mits_game', ['game_id'], ['id'], ondelete='CASCADE')
+    op.alter_column('mits_team', 'concurrent_attendees',
+               existing_type=sa.INTEGER(),
+               nullable=True)
+    op.create_unique_constraint(None, 'mits_times', ['team_id'])
+    op.drop_constraint(op.f('fk_mits_times_team_id_mits_team'), 'mits_times', type_='foreignkey')
+    op.create_foreign_key(None, 'mits_times', 'mits_team', ['team_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_no_shirt_attendee_id_attendee'), 'no_shirt', type_='foreignkey')
+    op.create_foreign_key(None, 'no_shirt', 'attendee', ['attendee_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_old_m_point_exchange_attendee_id_attendee'), 'old_m_point_exchange', type_='foreignkey')
+    op.create_foreign_key(None, 'old_m_point_exchange', 'attendee', ['attendee_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_panel_application_event_id_event'), 'panel_application', type_='foreignkey')
+    op.drop_constraint(op.f('fk_panel_application_poc_id_attendee'), 'panel_application', type_='foreignkey')
+    op.drop_constraint(op.f('fk_panel_application_submitter_id_panel_applicant'), 'panel_application', type_='foreignkey')
+    op.create_foreign_key(None, 'panel_application', 'attendee', ['poc_id'], ['id'])
+    op.create_foreign_key(None, 'panel_application', 'event', ['event_id'], ['id'])
+    op.create_foreign_key(None, 'panel_application', 'panel_applicant', ['submitter_id'], ['id'])
+    op.alter_column('password_reset', 'admin_id',
+               existing_type=sa.UUID(),
+               nullable=False)
+    op.alter_column('password_reset', 'attendee_id',
+               existing_type=sa.UUID(),
+               nullable=False)
+    op.drop_constraint(op.f('fk_password_reset_admin_id_admin_account'), 'password_reset', type_='foreignkey')
+    op.drop_constraint(op.f('fk_password_reset_attendee_id_attendee_account'), 'password_reset', type_='foreignkey')
+    op.create_foreign_key(None, 'password_reset', 'admin_account', ['admin_id'], ['id'], ondelete='CASCADE')
+    op.create_foreign_key(None, 'password_reset', 'attendee_account', ['attendee_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_print_queue_attendee_id_attendee'), 'print_job', type_='foreignkey')
+    op.create_foreign_key(None, 'print_job', 'attendee', ['attendee_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_promo_code_group_id_promo_code_group'), 'promo_code', type_='foreignkey')
+    op.create_foreign_key(None, 'promo_code', 'promo_code_group', ['group_id'], ['id'])
+    op.drop_constraint(op.f('fk_promo_code_group_buyer_id_attendee'), 'promo_code_group', type_='foreignkey')
+    op.create_foreign_key(None, 'promo_code_group', 'attendee', ['buyer_id'], ['id'])
+    op.drop_index(op.f('uq_promo_code_word_normalized_word_part_of_speech'), table_name='promo_code_word')
+    op.create_index('uq_promo_code_word_normalized_word_part_of_speech', 'promo_code_word', [sa.literal_column('lower(trim(word))'), 'part_of_speech'], unique=True)
+    op.alter_column('receipt_item', 'receipt_id',
+               existing_type=sa.UUID(),
+               nullable=False)
+    op.drop_constraint(op.f('fk_receipt_item_txn_id_receipt_transaction'), 'receipt_item', type_='foreignkey')
+    op.drop_constraint(op.f('fk_receipt_item_receipt_id_model_receipt'), 'receipt_item', type_='foreignkey')
+    op.create_foreign_key(None, 'receipt_item', 'receipt_transaction', ['txn_id'], ['id'])
+    op.create_foreign_key(None, 'receipt_item', 'model_receipt', ['receipt_id'], ['id'], ondelete='CASCADE')
+    op.alter_column('receipt_transaction', 'receipt_id',
+               existing_type=sa.UUID(),
+               nullable=False)
+    op.drop_constraint(op.f('fk_receipt_transaction_refunded_txn_id_receipt_transaction'), 'receipt_transaction', type_='foreignkey')
+    op.drop_constraint(op.f('fk_receipt_transaction_receipt_info_id_receipt_info'), 'receipt_transaction', type_='foreignkey')
+    op.drop_constraint(op.f('fk_receipt_transaction_receipt_id_model_receipt'), 'receipt_transaction', type_='foreignkey')
+    op.create_foreign_key(None, 'receipt_transaction', 'receipt_transaction', ['refunded_txn_id'], ['id'])
+    op.create_foreign_key(None, 'receipt_transaction', 'receipt_info', ['receipt_info_id'], ['id'])
+    op.create_foreign_key(None, 'receipt_transaction', 'model_receipt', ['receipt_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_room_assignment_attendee_id_attendee'), 'room_assignment', type_='foreignkey')
+    op.drop_constraint(op.f('fk_room_assignment_room_id_room'), 'room_assignment', type_='foreignkey')
+    op.create_foreign_key(None, 'room_assignment', 'attendee', ['attendee_id'], ['id'], ondelete='CASCADE')
+    op.create_foreign_key(None, 'room_assignment', 'room', ['room_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_sale_attendee_id_attendee'), 'sale', type_='foreignkey')
+    op.create_foreign_key(None, 'sale', 'attendee', ['attendee_id'], ['id'])
+    op.drop_constraint(op.f('fk_tabletop_checkout_game_id_tabletop_game'), 'tabletop_checkout', type_='foreignkey')
+    op.drop_constraint(op.f('fk_tabletop_checkout_attendee_id_attendee'), 'tabletop_checkout', type_='foreignkey')
+    op.create_foreign_key(None, 'tabletop_checkout', 'tabletop_game', ['game_id'], ['id'], ondelete='CASCADE')
+    op.create_foreign_key(None, 'tabletop_checkout', 'attendee', ['attendee_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint(op.f('fk_tabletop_game_attendee_id_attendee'), 'tabletop_game', type_='foreignkey')
+    op.create_foreign_key(None, 'tabletop_game', 'attendee', ['attendee_id'], ['id'], ondelete='CASCADE')
+    op.drop_index(op.f('ix_tracking_who'), table_name='tracking')
+    op.drop_constraint(op.f('txn_request_tracking_incr_id_key'), 'txn_request_tracking', type_='unique')
+
+
+def downgrade():
+    op.create_unique_constraint(op.f('txn_request_tracking_incr_id_key'), 'txn_request_tracking', ['incr_id'], postgresql_nulls_not_distinct=False)
+    op.create_index(op.f('ix_tracking_who'), 'tracking', ['who'], unique=False)
+    op.drop_constraint(None, 'tabletop_game', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_tabletop_game_attendee_id_attendee'), 'tabletop_game', 'attendee', ['attendee_id'], ['id'])
+    op.drop_constraint(None, 'tabletop_checkout', type_='foreignkey')
+    op.drop_constraint(None, 'tabletop_checkout', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_tabletop_checkout_attendee_id_attendee'), 'tabletop_checkout', 'attendee', ['attendee_id'], ['id'])
+    op.create_foreign_key(op.f('fk_tabletop_checkout_game_id_tabletop_game'), 'tabletop_checkout', 'tabletop_game', ['game_id'], ['id'])
+    op.drop_constraint(None, 'sale', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_sale_attendee_id_attendee'), 'sale', 'attendee', ['attendee_id'], ['id'], ondelete='SET NULL')
+    op.drop_constraint(None, 'room_assignment', type_='foreignkey')
+    op.drop_constraint(None, 'room_assignment', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_room_assignment_room_id_room'), 'room_assignment', 'room', ['room_id'], ['id'])
+    op.create_foreign_key(op.f('fk_room_assignment_attendee_id_attendee'), 'room_assignment', 'attendee', ['attendee_id'], ['id'])
+    op.drop_constraint(None, 'receipt_transaction', type_='foreignkey')
+    op.drop_constraint(None, 'receipt_transaction', type_='foreignkey')
+    op.drop_constraint(None, 'receipt_transaction', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_receipt_transaction_receipt_id_model_receipt'), 'receipt_transaction', 'model_receipt', ['receipt_id'], ['id'], ondelete='SET NULL')
+    op.create_foreign_key(op.f('fk_receipt_transaction_receipt_info_id_receipt_info'), 'receipt_transaction', 'receipt_info', ['receipt_info_id'], ['id'], ondelete='SET NULL')
+    op.create_foreign_key(op.f('fk_receipt_transaction_refunded_txn_id_receipt_transaction'), 'receipt_transaction', 'receipt_transaction', ['refunded_txn_id'], ['id'], ondelete='SET NULL')
+    op.alter_column('receipt_transaction', 'receipt_id',
+               existing_type=sa.UUID(),
+               nullable=True)
+    op.drop_constraint(None, 'receipt_item', type_='foreignkey')
+    op.drop_constraint(None, 'receipt_item', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_receipt_item_receipt_id_model_receipt'), 'receipt_item', 'model_receipt', ['receipt_id'], ['id'], ondelete='SET NULL')
+    op.create_foreign_key(op.f('fk_receipt_item_txn_id_receipt_transaction'), 'receipt_item', 'receipt_transaction', ['txn_id'], ['id'], ondelete='SET NULL')
+    op.alter_column('receipt_item', 'receipt_id',
+               existing_type=sa.UUID(),
+               nullable=True)
+    op.drop_index('uq_promo_code_word_normalized_word_part_of_speech', table_name='promo_code_word')
+    op.create_index(op.f('uq_promo_code_word_normalized_word_part_of_speech'), 'promo_code_word', [sa.literal_column('lower(TRIM(BOTH FROM word))'), 'part_of_speech'], unique=True)
+    op.drop_constraint(None, 'promo_code_group', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_promo_code_group_buyer_id_attendee'), 'promo_code_group', 'attendee', ['buyer_id'], ['id'], ondelete='SET NULL')
+    op.drop_constraint(None, 'promo_code', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_promo_code_group_id_promo_code_group'), 'promo_code', 'promo_code_group', ['group_id'], ['id'], ondelete='SET NULL')
+    op.drop_constraint(None, 'print_job', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_print_queue_attendee_id_attendee'), 'print_job', 'attendee', ['attendee_id'], ['id'])
+    op.drop_constraint(None, 'password_reset', type_='foreignkey')
+    op.drop_constraint(None, 'password_reset', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_password_reset_attendee_id_attendee_account'), 'password_reset', 'attendee_account', ['attendee_id'], ['id'])
+    op.create_foreign_key(op.f('fk_password_reset_admin_id_admin_account'), 'password_reset', 'admin_account', ['admin_id'], ['id'])
+    op.alter_column('password_reset', 'attendee_id',
+               existing_type=sa.UUID(),
+               nullable=True)
+    op.alter_column('password_reset', 'admin_id',
+               existing_type=sa.UUID(),
+               nullable=True)
+    op.drop_constraint(None, 'panel_application', type_='foreignkey')
+    op.drop_constraint(None, 'panel_application', type_='foreignkey')
+    op.drop_constraint(None, 'panel_application', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_panel_application_submitter_id_panel_applicant'), 'panel_application', 'panel_applicant', ['submitter_id'], ['id'], ondelete='SET NULL')
+    op.create_foreign_key(op.f('fk_panel_application_poc_id_attendee'), 'panel_application', 'attendee', ['poc_id'], ['id'], ondelete='SET NULL')
+    op.create_foreign_key(op.f('fk_panel_application_event_id_event'), 'panel_application', 'event', ['event_id'], ['id'], ondelete='SET NULL')
+    op.drop_constraint(None, 'old_m_point_exchange', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_old_m_point_exchange_attendee_id_attendee'), 'old_m_point_exchange', 'attendee', ['attendee_id'], ['id'])
+    op.drop_constraint(None, 'no_shirt', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_no_shirt_attendee_id_attendee'), 'no_shirt', 'attendee', ['attendee_id'], ['id'])
+    op.drop_constraint(None, 'mits_times', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_mits_times_team_id_mits_team'), 'mits_times', 'mits_team', ['team_id'], ['id'])
+    op.drop_constraint(None, 'mits_times', type_='unique')
+    op.alter_column('mits_team', 'concurrent_attendees',
+               existing_type=sa.INTEGER(),
+               nullable=False)
+    op.drop_constraint(None, 'mits_picture', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_mits_picture_game_id_mits_game'), 'mits_picture', 'mits_game', ['game_id'], ['id'])
+    op.drop_constraint(None, 'mits_panel_application', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_mits_panel_application_team_id_mits_team'), 'mits_panel_application', 'mits_team', ['team_id'], ['id'])
+    op.drop_constraint(None, 'mits_panel_application', type_='unique')
+    op.drop_constraint(None, 'mits_game', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_mits_game_team_id_mits_team'), 'mits_game', 'mits_team', ['team_id'], ['id'])
+    op.drop_constraint(None, 'mits_document', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_mits_document_game_id_mits_game'), 'mits_document', 'mits_game', ['game_id'], ['id'])
+    op.drop_constraint(None, 'mits_applicant', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_mits_applicant_team_id_mits_team'), 'mits_applicant', 'mits_team', ['team_id'], ['id'])
+    op.alter_column('merch_pickup', 'picked_up_for_id',
+               existing_type=sa.UUID(),
+               nullable=False)
+    op.alter_column('merch_pickup', 'picked_up_by_id',
+               existing_type=sa.UUID(),
+               nullable=False)
+    op.drop_constraint(None, 'merch_discount', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_merch_discount_attendee_id_attendee'), 'merch_discount', 'attendee', ['attendee_id'], ['id'])
+    op.drop_constraint(None, 'm_points_for_cash', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_m_points_for_cash_attendee_id_attendee'), 'm_points_for_cash', 'attendee', ['attendee_id'], ['id'])
+    op.create_unique_constraint(op.f('uq_lottery_application_response_id'), 'lottery_application', ['response_id'], postgresql_nulls_not_distinct=False)
+    op.drop_constraint(None, 'job_template', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_job_template_department_id_department'), 'job_template', 'department', ['department_id'], ['id'])
+    op.drop_constraint(None, 'job', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_job_department_id_department'), 'job', 'department', ['department_id'], ['id'])
+    op.drop_constraint(None, 'indie_studio', type_='unique')
+    op.create_unique_constraint(op.f('uq_indie_studio_name'), 'indie_studio', ['name'], postgresql_nulls_not_distinct=False)
+    op.drop_constraint(None, 'indie_judge', type_='unique')
+    op.alter_column('indie_judge', 'admin_id',
+               existing_type=sa.UUID(),
+               nullable=False)
+    op.drop_constraint(None, 'indie_game_review', type_='foreignkey')
+    op.drop_constraint(None, 'indie_game_review', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_indie_game_review_judge_id_indie_judge'), 'indie_game_review', 'indie_judge', ['judge_id'], ['id'])
+    op.create_foreign_key(op.f('fk_indie_game_review_game_id_indie_game'), 'indie_game_review', 'indie_game', ['game_id'], ['id'])
+    op.drop_constraint(None, 'indie_game_image', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_indie_game_image_game_id_indie_game'), 'indie_game_image', 'indie_game', ['game_id'], ['id'])
+    op.drop_constraint(None, 'indie_game_code', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_indie_game_code_game_id_indie_game'), 'indie_game_code', 'indie_game', ['game_id'], ['id'])
+    op.drop_constraint(None, 'indie_game', type_='foreignkey')
+    op.drop_constraint(None, 'indie_game', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_indie_game_studio_id_indie_studio'), 'indie_game', 'indie_studio', ['studio_id'], ['id'])
+    op.create_foreign_key(op.f('fk_indie_game_primary_contact_id_indie_developer'), 'indie_game', 'indie_developer', ['primary_contact_id'], ['id'], ondelete='SET NULL')
+    op.alter_column('indie_game', 'multiplayer_game_length',
+               existing_type=sa.INTEGER(),
+               nullable=True)
+    op.drop_constraint(None, 'indie_developer', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_indie_developer_studio_id_indie_studio'), 'indie_developer', 'indie_studio', ['studio_id'], ['id'])
+    op.drop_constraint(None, 'hotel_requests', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_hotel_requests_attendee_id_attendee'), 'hotel_requests', 'attendee', ['attendee_id'], ['id'])
+    op.drop_constraint(None, 'guest_travel_plans', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_guest_travel_plans_guest_id_guest_group'), 'guest_travel_plans', 'guest_group', ['guest_id'], ['id'])
+    op.drop_constraint(None, 'guest_track', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_guest_track_guest_id_guest_group'), 'guest_track', 'guest_group', ['guest_id'], ['id'])
+    op.drop_constraint(None, 'guest_track', type_='unique')
+    op.drop_constraint(None, 'guest_taxes', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_guest_taxes_guest_id_guest_group'), 'guest_taxes', 'guest_group', ['guest_id'], ['id'])
+    op.drop_constraint(None, 'guest_stage_plot', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_guest_stage_plot_guest_id_guest_group'), 'guest_stage_plot', 'guest_group', ['guest_id'], ['id'])
+    op.drop_constraint(None, 'guest_panel', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_guest_panel_guest_id_guest_group'), 'guest_panel', 'guest_group', ['guest_id'], ['id'])
+    op.drop_constraint(None, 'guest_merch', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_guest_merch_guest_id_guest_group'), 'guest_merch', 'guest_group', ['guest_id'], ['id'])
+    op.drop_constraint(None, 'guest_media_request', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_guest_media_request_guest_id_guest_group'), 'guest_media_request', 'guest_group', ['guest_id'], ['id'])
+    op.drop_constraint(None, 'guest_interview', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_guest_interview_guest_id_guest_group'), 'guest_interview', 'guest_group', ['guest_id'], ['id'])
+    op.drop_constraint(None, 'guest_info', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_guest_info_guest_id_guest_group'), 'guest_info', 'guest_group', ['guest_id'], ['id'])
+    op.drop_constraint(None, 'guest_image', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_guest_image_guest_id_guest_group'), 'guest_image', 'guest_group', ['guest_id'], ['id'])
+    op.drop_constraint(None, 'guest_hospitality', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_guest_hospitality_guest_id_guest_group'), 'guest_hospitality', 'guest_group', ['guest_id'], ['id'])
+    op.drop_constraint(None, 'guest_group', type_='foreignkey')
+    op.drop_constraint(None, 'guest_group', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_guest_group_event_id_event'), 'guest_group', 'event', ['event_id'], ['id'], ondelete='SET NULL')
+    op.create_foreign_key(op.f('fk_guest_group_group_id_group'), 'guest_group', 'group', ['group_id'], ['id'])
+    op.drop_constraint(None, 'guest_group', type_='unique')
+    op.drop_constraint(None, 'guest_detailed_travel_plan', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_guest_detailed_travel_plan_travel_plans_id_guest_tra_6ad4'), 'guest_detailed_travel_plan', 'guest_travel_plans', ['travel_plans_id'], ['id'])
+    op.alter_column('guest_detailed_travel_plan', 'travel_plans_id',
+               existing_type=sa.UUID(),
+               nullable=True)
+    op.drop_constraint(None, 'guest_charity', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_guest_charity_guest_id_guest_group'), 'guest_charity', 'guest_group', ['guest_id'], ['id'])
+    op.drop_constraint(None, 'guest_bio', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_guest_bio_guest_id_guest_group'), 'guest_bio', 'guest_group', ['guest_id'], ['id'])
+    op.drop_constraint(None, 'guest_autograph', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_guest_autograph_guest_id_guest_group'), 'guest_autograph', 'guest_group', ['guest_id'], ['id'])
+    op.drop_constraint(None, 'group', type_='foreignkey')
+    op.drop_constraint(None, 'group', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_group_shared_with_id_group'), 'group', 'group', ['shared_with_id'], ['id'], ondelete='SET NULL')
+    op.create_foreign_key(op.f('fk_leader'), 'group', 'attendee', ['leader_id'], ['id'], ondelete='SET NULL')
+    op.drop_constraint(None, 'food_restrictions', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_food_restrictions_attendee_id_attendee'), 'food_restrictions', 'attendee', ['attendee_id'], ['id'])
+    op.drop_constraint(None, 'event_location', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_event_location_department_id_department'), 'event_location', 'department', ['department_id'], ['id'], ondelete='SET NULL')
+    op.drop_constraint(None, 'event_feedback', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_event_feedback_event_id_event'), 'event_feedback', 'event', ['event_id'], ['id'])
+    op.drop_constraint(None, 'event', type_='foreignkey')
+    op.drop_constraint(None, 'event', type_='foreignkey')
+    op.drop_constraint(None, 'event', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_event_location_id_event_location'), 'event', 'event_location', ['event_location_id'], ['id'], ondelete='SET NULL')
+    op.create_foreign_key(op.f('fk_event_department_id_department'), 'event', 'department', ['department_id'], ['id'], ondelete='SET NULL')
+    op.create_foreign_key(op.f('fk_event_attraction_event_id_attraction_event'), 'event', 'attraction_event', ['attraction_event_id'], ['id'], ondelete='SET NULL')
+    op.drop_constraint(None, 'event', type_='unique')
+    op.drop_constraint(None, 'email', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_email_automated_email_id_automated_email'), 'email', 'automated_email', ['automated_email_id'], ['id'], ondelete='SET NULL')
+    op.drop_constraint(None, 'dept_role', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_dept_role_department_id_department'), 'dept_role', 'department', ['department_id'], ['id'])
+    op.drop_constraint(None, 'dept_membership_request', type_='foreignkey')
+    op.drop_constraint(None, 'dept_membership_request', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_dept_membership_request_department_id_department'), 'dept_membership_request', 'department', ['department_id'], ['id'])
+    op.create_foreign_key(op.f('fk_dept_membership_request_attendee_id_attendee'), 'dept_membership_request', 'attendee', ['attendee_id'], ['id'])
+    op.drop_constraint(None, 'dept_membership', type_='foreignkey')
+    op.drop_constraint(None, 'dept_membership', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_dept_membership_attendee_id_attendee'), 'dept_membership', 'attendee', ['attendee_id'], ['id'])
+    op.create_foreign_key(op.f('fk_dept_membership_department_id_department'), 'dept_membership', 'department', ['department_id'], ['id'])
+    op.drop_constraint(None, 'dept_checklist_item', type_='foreignkey')
+    op.drop_constraint(None, 'dept_checklist_item', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_dept_checklist_item_attendee_id_attendee'), 'dept_checklist_item', 'attendee', ['attendee_id'], ['id'])
+    op.create_foreign_key(op.f('fk_dept_checklist_item_department_id_department'), 'dept_checklist_item', 'department', ['department_id'], ['id'])
+    op.drop_constraint(None, 'department', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_department_parent_id_department'), 'department', 'department', ['parent_id'], ['id'])
+    op.drop_constraint(None, 'bulk_printing_request', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_bulk_printing_request_department_id_department'), 'bulk_printing_request', 'department', ['department_id'], ['id'])
+    op.drop_constraint(None, 'badge_info', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_badge_info_attendee_id_attendee'), 'badge_info', 'attendee', ['attendee_id'], ['id'], ondelete='SET NULL')
+    op.create_index(op.f('ix_badge_info_ident'), 'badge_info', ['ident'], unique=False)
+    op.create_unique_constraint(op.f('uq_automated_email_ident'), 'automated_email', ['ident'], postgresql_nulls_not_distinct=False)
+    op.drop_constraint(None, 'attraction_signup', type_='foreignkey')
+    op.drop_constraint(None, 'attraction_signup', type_='foreignkey')
+    op.drop_constraint(None, 'attraction_signup', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_attraction_signup_attraction_event_id_attraction_event'), 'attraction_signup', 'attraction_event', ['attraction_event_id'], ['id'])
+    op.create_foreign_key(op.f('fk_attraction_signup_attraction_id_attraction'), 'attraction_signup', 'attraction', ['attraction_id'], ['id'])
+    op.create_foreign_key(op.f('fk_attraction_signup_attendee_id_attendee'), 'attraction_signup', 'attendee', ['attendee_id'], ['id'])
+    op.create_index(op.f('ix_attraction_notification_reply_sid'), 'attraction_notification_reply', ['sid'], unique=False)
+    op.drop_constraint(None, 'attraction_notification', type_='foreignkey')
+    op.drop_constraint(None, 'attraction_notification', type_='foreignkey')
+    op.drop_constraint(None, 'attraction_notification', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_attraction_notification_attendee_id_attendee'), 'attraction_notification', 'attendee', ['attendee_id'], ['id'])
+    op.create_foreign_key(op.f('fk_attraction_notification_attraction_event_id_attraction_event'), 'attraction_notification', 'attraction_event', ['attraction_event_id'], ['id'])
+    op.create_foreign_key(op.f('fk_attraction_notification_attraction_id_attraction'), 'attraction_notification', 'attraction', ['attraction_id'], ['id'])
+    op.create_index(op.f('ix_attraction_notification_ident'), 'attraction_notification', ['ident'], unique=False)
+    op.drop_constraint(None, 'attraction_feature', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_attraction_feature_attraction_id_attraction'), 'attraction_feature', 'attraction', ['attraction_id'], ['id'])
+    op.drop_constraint(None, 'attraction_event', type_='foreignkey')
+    op.drop_constraint(None, 'attraction_event', type_='foreignkey')
+    op.drop_constraint(None, 'attraction_event', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_attraction_event_location_id_event_location'), 'attraction_event', 'event_location', ['event_location_id'], ['id'], ondelete='SET NULL')
+    op.create_foreign_key(op.f('fk_attraction_event_attraction_id_attraction'), 'attraction_event', 'attraction', ['attraction_id'], ['id'])
+    op.create_foreign_key(op.f('fk_attraction_event_attraction_feature_id_attraction_feature'), 'attraction_event', 'attraction_feature', ['attraction_feature_id'], ['id'])
+    op.create_unique_constraint(op.f('uq_attraction_slug'), 'attraction', ['slug'], postgresql_nulls_not_distinct=False)
+    op.create_unique_constraint(op.f('attraction_name_key'), 'attraction', ['name'], postgresql_nulls_not_distinct=False)
+    op.add_column('attendee', sa.Column('affiliate', sa.VARCHAR(), server_default=sa.text("''::character varying"), autoincrement=False, nullable=False))
+    op.drop_constraint(None, 'attendee', type_='foreignkey')
+    op.drop_constraint(None, 'attendee', type_='foreignkey')
+    op.drop_constraint(None, 'attendee', type_='foreignkey')
+    op.drop_constraint(None, 'attendee', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_attendee_badge_pickup_group_id_badge_pickup_group'), 'attendee', 'badge_pickup_group', ['badge_pickup_group_id'], ['id'], ondelete='SET NULL')
+    op.create_foreign_key(op.f('fk_attendee_group_id_group'), 'attendee', 'group', ['group_id'], ['id'], ondelete='SET NULL')
+    op.create_foreign_key(op.f('fk_attendee_watchlist_id_watch_list'), 'attendee', 'watch_list', ['watchlist_id'], ['id'], ondelete='SET NULL')
+    op.create_foreign_key(op.f('fk_attendee_creator_id_attendee'), 'attendee', 'attendee', ['creator_id'], ['id'], ondelete='SET NULL')
+    op.drop_constraint(None, 'artist_marketplace_application', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_artist_marketplace_application_attendee_id_attendee'), 'artist_marketplace_application', 'attendee', ['attendee_id'], ['id'])
+    op.drop_constraint(None, 'artist_marketplace_application', type_='unique')
+    op.drop_constraint(None, 'art_show_receipt', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_art_show_receipt_attendee_id_attendee'), 'art_show_receipt', 'attendee', ['attendee_id'], ['id'], ondelete='SET NULL')
+    op.alter_column('art_show_receipt', 'attendee_id',
+               existing_type=sa.UUID(),
+               nullable=True)
+    op.drop_constraint(None, 'art_show_piece', type_='foreignkey')
+    op.drop_constraint(None, 'art_show_piece', type_='foreignkey')
+    op.drop_constraint(None, 'art_show_piece', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_art_show_pieces_app_id_art_show_application'), 'art_show_piece', 'art_show_application', ['app_id'], ['id'], ondelete='SET NULL')
+    op.create_foreign_key(op.f('fk_art_show_piece_winning_bidder_id_art_show_bidder'), 'art_show_piece', 'art_show_bidder', ['winning_bidder_id'], ['id'], ondelete='SET NULL')
+    op.create_foreign_key(op.f('fk_art_show_piece_receipt_id_art_show_receipt'), 'art_show_piece', 'art_show_receipt', ['receipt_id'], ['id'], ondelete='SET NULL')
+    op.alter_column('art_show_piece', 'app_id',
+               existing_type=sa.UUID(),
+               nullable=True)
+    op.drop_constraint(None, 'art_show_payment', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_art_show_payment_receipt_id_art_show_receipt'), 'art_show_payment', 'art_show_receipt', ['receipt_id'], ['id'], ondelete='SET NULL')
+    op.alter_column('art_show_payment', 'receipt_id',
+               existing_type=sa.UUID(),
+               nullable=True)
+    op.drop_constraint(None, 'art_show_bidder', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_art_show_bidder_attendee_id_attendee'), 'art_show_bidder', 'attendee', ['attendee_id'], ['id'], ondelete='SET NULL')
+    op.alter_column('art_show_bidder', 'attendee_id',
+               existing_type=sa.UUID(),
+               nullable=True)
+    op.drop_constraint(None, 'art_show_application', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_art_show_application_attendee_id_attendee'), 'art_show_application', 'attendee', ['attendee_id'], ['id'], ondelete='SET NULL')
+    op.drop_constraint(None, 'art_show_application', type_='unique')
+    op.alter_column('art_show_application', 'default_cost',
+               existing_type=sa.INTEGER(),
+               nullable=True)
+    op.alter_column('art_show_application', 'overridden_price',
+               existing_type=sa.INTEGER(),
+               nullable=True)
+    op.drop_constraint(None, 'art_show_agent_code', type_='foreignkey')
+    op.drop_constraint(None, 'art_show_agent_code', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_art_show_agent_code_app_id_art_show_application'), 'art_show_agent_code', 'art_show_application', ['app_id'], ['id'])
+    op.create_foreign_key(op.f('fk_art_show_agent_code_attendee_id_attendee'), 'art_show_agent_code', 'attendee', ['attendee_id'], ['id'], ondelete='SET NULL')
+    op.drop_constraint(None, 'art_panel_assignment', type_='foreignkey')
+    op.drop_constraint(None, 'art_panel_assignment', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_art_panel_assignment_app_id_art_show_application'), 'art_panel_assignment', 'art_show_application', ['app_id'], ['id'])
+    op.create_foreign_key(op.f('fk_art_panel_assignment_panel_id_art_show_panel'), 'art_panel_assignment', 'art_show_panel', ['panel_id'], ['id'])
+    op.drop_constraint(None, 'api_token', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_api_token_admin_account_id_admin_account'), 'api_token', 'admin_account', ['admin_account_id'], ['id'])
+    op.drop_constraint(None, 'admin_account', type_='foreignkey')
+    op.create_foreign_key(op.f('fk_admin_account_attendee_id_attendee'), 'admin_account', 'attendee', ['attendee_id'], ['id'])
+    op.create_table('attendee_tournament',
+    sa.Column('id', sa.UUID(), autoincrement=False, nullable=False),
+    sa.Column('first_name', sa.VARCHAR(), server_default=sa.text("''::character varying"), autoincrement=False, nullable=False),
+    sa.Column('last_name', sa.VARCHAR(), server_default=sa.text("''::character varying"), autoincrement=False, nullable=False),
+    sa.Column('email', sa.VARCHAR(), server_default=sa.text("''::character varying"), autoincrement=False, nullable=False),
+    sa.Column('cellphone', sa.VARCHAR(), server_default=sa.text("''::character varying"), autoincrement=False, nullable=False),
+    sa.Column('game', sa.VARCHAR(), server_default=sa.text("''::character varying"), autoincrement=False, nullable=False),
+    sa.Column('availability', sa.VARCHAR(), server_default=sa.text("''::character varying"), autoincrement=False, nullable=False),
+    sa.Column('format', sa.VARCHAR(), server_default=sa.text("''::character varying"), autoincrement=False, nullable=False),
+    sa.Column('experience', sa.VARCHAR(), server_default=sa.text("''::character varying"), autoincrement=False, nullable=False),
+    sa.Column('needs', sa.VARCHAR(), server_default=sa.text("''::character varying"), autoincrement=False, nullable=False),
+    sa.Column('why', sa.VARCHAR(), server_default=sa.text("''::character varying"), autoincrement=False, nullable=False),
+    sa.Column('volunteering', sa.BOOLEAN(), server_default=sa.text('false'), autoincrement=False, nullable=False),
+    sa.Column('status', sa.INTEGER(), server_default=sa.text('239694250'), autoincrement=False, nullable=False),
+    sa.Column('created', postgresql.TIMESTAMP(timezone=True), server_default=sa.text("timezone('utc'::text, CURRENT_TIMESTAMP)"), autoincrement=False, nullable=False),
+    sa.Column('last_updated', postgresql.TIMESTAMP(timezone=True), server_default=sa.text("timezone('utc'::text, CURRENT_TIMESTAMP)"), autoincrement=False, nullable=False),
+    sa.Column('external_id', postgresql.JSONB(astext_type=sa.Text()), server_default=sa.text("'{}'::jsonb"), autoincrement=False, nullable=False),
+    sa.Column('last_synced', postgresql.JSONB(astext_type=sa.Text()), server_default=sa.text("'{}'::jsonb"), autoincrement=False, nullable=False),
+    sa.PrimaryKeyConstraint('id', name=op.f('pk_attendee_tournament'))
+    )
