+++ uber/api.py
@@ -1699,7 +1699,7 @@ class PrintJobLookup:
                 if not restart or not errors:
                     results[job.id] = self._build_job_json_data(job)
                     if not dry_run:
-                        job.queued = datetime.utcnow()
+                        job.queued = datetime.now(UTC)
                         session.add(job)
                         session.commit()
 
@@ -1793,7 +1793,7 @@ class PrintJobLookup:
 
             for job in jobs:
                 results[job.id] = self._build_job_json_data(job)
-                job.printed = datetime.utcnow()
+                job.printed = datetime.now(UTC)
                 session.add(job)
                 session.commit()
 
@@ -1839,7 +1839,7 @@ class PrintJobLookup:
                     else:
                         job.errors = error
                 else:
-                    job.printed = datetime.utcnow()
+                    job.printed = datetime.now(UTC)
                 session.add(job)
                 session.commit()
