+++ uber/models/api.py
@@ -5,12 +5,11 @@ from pytz import UTC
 from sqlalchemy import String, Uuid, DateTime
 from sqlalchemy.dialects.postgresql.json import JSONB
 from sqlalchemy.ext.mutable import MutableDict
-from sqlmodel import Field, Relationship
 from typing import Any
 
 from uber.config import c
 from uber.models import MagModel
-from uber.models.types import DefaultColumn as Column, MultiChoice
+from uber.models.types import DefaultColumn as Column, MultiChoice, DefaultField as Field, DefaultRelationship as Relationship
 
 
 __all__ = ['ApiToken', 'ApiJob']
@@ -24,12 +23,12 @@ class ApiToken(MagModel, table=True):
     admin_account_id: str | None = Field(sa_type=Uuid(as_uuid=False), foreign_key='admin_account.id', ondelete='CASCADE')
     admin_account: "AdminAccount" = Relationship(back_populates="api_tokens", sa_relationship_kwargs={'lazy': 'joined'})
 
-    token: str | None = Column(Uuid(as_uuid=False), default=lambda: str(uuid.uuid4()), private=True)
-    access: str = Column(MultiChoice(c.API_ACCESS_OPTS))
-    name: str = Column(String)
-    description: str = Column(String)
-    issued_time: datetime = Column(DateTime(timezone=True), default=lambda: datetime.now(UTC))
-    revoked_time: datetime = Column(DateTime(timezone=True), default=None, nullable=True)
+    token: str | None = Field(sa_type=Uuid(as_uuid=False), default_factory=lambda: str(uuid.uuid4()), private=True)
+    access: str = Field(sa_column=Column(MultiChoice(c.API_ACCESS_OPTS)))
+    name: str = ''
+    description: str = ''
+    issued_time: datetime = Field(sa_type=DateTime(timezone=True), default_factory=lambda: datetime.now(UTC))
+    revoked_time: datetime = Field(sa_type=DateTime(timezone=True), default=None, nullable=True)
 
     @property
     def api_read(self):
@@ -52,13 +51,13 @@ class ApiJob(MagModel, table=True):
     admin_id: str | None = Field(sa_type=Uuid(as_uuid=False), foreign_key='admin_account.id', nullable=True)
     admin_account: "AdminAccount" = Relationship(back_populates="api_jobs")
 
-    admin_name: str = Column(String)  # Preserve admin's name in case their account is removed
-    queued: datetime | None = Column(DateTime(timezone=True), nullable=True, default=None)
-    completed: datetime | None = Column(DateTime(timezone=True), nullable=True, default=None)
-    cancelled: datetime | None = Column(DateTime(timezone=True), nullable=True, default=None)
-    job_name: str = Column(String)
-    target_server: str = Column(String)
-    query: str = Column(String)
-    api_token: str = Column(String)
-    errors: str = Column(String)
+    admin_name: str = ''  # Preserve admin's name in case their account is removed
+    queued: datetime | None = Field(sa_type=DateTime(timezone=True), nullable=True, default=None)
+    completed: datetime | None = Field(sa_type=DateTime(timezone=True), nullable=True, default=None)
+    cancelled: datetime | None = Field(sa_type=DateTime(timezone=True), nullable=True, default=None)
+    job_name: str = ''
+    target_server: str = ''
+    query: str = ''
+    api_token: str = ''
+    errors: str = ''
     json_data: dict[str, Any] = Field(sa_type=MutableDict.as_mutable(JSONB), default_factory=dict)
