+++ uber/models/legal.py
@@ -1,23 +1,22 @@
 from datetime import datetime
 from sqlalchemy.types import Uuid, String, DateTime
-from sqlmodel import Field, Relationship
 
 from uber.decorators import presave_adjustment
 from uber.models import MagModel
-from uber.models.types import DefaultColumn as Column
+from uber.models.types import DefaultColumn as Column, DefaultField as Field, DefaultRelationship as Relationship
 
 __all__ = ['SignedDocument']
 
 
 class SignedDocument(MagModel, table=True):
-    fk_id: str = Column(Uuid(as_uuid=False), index=True)
+    fk_id: str = Field(sa_type=Uuid(as_uuid=False), index=True)
     model: str = Column(String)
     document_id: str = Column(String)
-    last_emailed: datetime | None = Column(DateTime(timezone=True), nullable=True, default=None)
+    last_emailed: datetime | None = Field(sa_type=DateTime(timezone=True), nullable=True, default=None)
     link: str = Column(String)
     ident: str = Column(String)
-    signed: datetime | None = Column(DateTime(timezone=True), nullable=True, default=None)
-    declined: datetime | None = Column(DateTime(timezone=True), nullable=True, default=None)
+    signed: datetime | None = Field(sa_type=DateTime(timezone=True), nullable=True, default=None)
+    declined: datetime | None = Field(sa_type=DateTime(timezone=True), nullable=True, default=None)
 
     @presave_adjustment
     def null_to_strings(self):
