+++ uber/site_sections/devtools.py
@@ -234,7 +234,7 @@ class Root:
         db_read_time += time.perf_counter()
 
         return json.dumps({
-            'server_current_timestamp': int(datetime.utcnow().timestamp()),
+            'server_current_timestamp': int(datetime.now(UTC).timestamp()),
             'session_read_count': read_count,
             'session_commit_time': session_commit_time,
             'db_read_time': db_read_time,
